scheduler enhancement: Optimize the code.
